Fix recursive CTE regression and update tests
Address regression in recursive CTE handling by replacing the strict
projection with a specific alignment path. This ensures proper column
alignment, maintains nullability, and validates type compatibility,
while preserving the output schema contract for planner checks.
Move helper tests to common.rs and retain integration coverage in
recursive_query.rs. Restore behavior in cte.slt.
